+++ trunk/macro.h
@@ -31,7 +31,7 @@ struct Mfunc;
 extern bool mtranslate(struct file_s *);
 extern Obj *macro_recurse(Wait_types, Obj *, struct Namespace *, linepos_t);
 extern Obj *mfunc_recurse(Wait_types, struct Mfunc *, struct Namespace *, linepos_t, uint8_t);
-extern Obj *mfunc2_recurse(struct Mfunc *, struct values_s *, unsigned int, linepos_t);
+extern Obj *mfunc2_recurse(struct Mfunc *, struct values_s *, size_t, linepos_t);
 extern void init_macro(void);
 extern void free_macro(void);
 extern void get_macro_params(Obj *);
